+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+ - `test_type_error` to `TestVerifyIdentity`
+
+### Changed
+ - `test_error` to `test_nvi_exception`
+ - `test_nvi_exception`'s target exception to `NVIException` from plain
+ `Exception`
+
 ## [0.1.0pre5] - 2019-03-13
 ### Changed
  - Moved package docstring to the top.